booking-service/requirements.txt
fastapi>=0.110.0,<0.111.0
uvicorn[standard]>=0.29.0,<0.30.0
pydantic>=2.6.0,<2.7.0
cassandra-driver>=3.29.0,<3.30.0
redis>=5.0.3,<5.1.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx>=0.27.0,<0.28.0
python-multipart==0.0.9
pika>=1.3.2,<1.4.0 # For RabbitMQ if used directly, though notification.py uses HTTP
python-consul2==0.1.5 # Switched to python-consul2
uuid==1.30 # For UUID generation, though built-in uuid is usually sufficient

# Ensure python-consul is removed if python-consul2 is used.
# pip uninstall python-consul